Seperated side panel from doctorprofile.jsx
